+++ saleor/discount/utils/voucher.py
@@ -72,8 +72,7 @@ def increase_voucher_usage(
     customer_email: str | None,
     increase_voucher_customer_usage: bool = True,
 ) -> None:
-    if voucher.usage_limit:
-        increase_voucher_code_usage_value(code)
+    increase_voucher_code_usage_value(code)
     if voucher.apply_once_per_customer and increase_voucher_customer_usage:
         add_voucher_usage_by_customer(code, customer_email)
     if voucher.single_use:
@@ -137,8 +136,7 @@ def release_voucher_code_usage(
 ):
     if not code:
         return
-    if voucher and voucher.usage_limit:
-        decrease_voucher_code_usage_value(code)
+    decrease_voucher_code_usage_value(code)
     if voucher and voucher.single_use:
         activate_voucher_code(code)
     if user_email: